factor out base config
