+++ packages/modular-scripts/react-scripts/config/parts/styleLoadersConfig.js
